create Concerns::TwitterUser::Api
