+++ app/models/concerns/twitter_user/utils.rb
@@ -31,5 +31,17 @@ module Concerns::TwitterUser::Utils
     def latest_me
       TwitterUser.latest(__uid_i, user_id)
     end
+
+    def search_and_touch
+      update!(search_count: search_count + 1)
+    rescue => e
+      logger.error "#{self.class}##{__method__} #{e.class} #{e.message}"
+    end
+
+    def update_and_touch
+      update!(update_count: update_count + 1)
+    rescue => e
+      logger.error "#{self.class}##{__method__} #{e.class} #{e.message}"
+    end
   end
 end
